requirements.txt
alembic==1.13.3
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cloudinary
cryptography==43.0.1
dnspython==2.7.0
eventlet==0.39.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
packaging==24.1
passlib==1.7.4
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
validators==0.34.0
Werkzeug==3.0.4
wsproto==1.2.0
zipp==3.20.2
